requirements.txt
django
gunicorn
djangorestframework
drf-spectacular
django-import-export
